refactor: extract progress bar stuff to its own file
